+++ src/navigation/router.js
@@ -1,42 +1,28 @@
 // @flow
 import React, { Component } from 'react'
-import { Router, Route, IndexRoute, browserHistory, applyRouterMiddleware } from 'react-router'
+import { Router, browserHistory, applyRouterMiddleware } from 'react-router'
 import useRelay from 'react-router-relay'
 import useScroll from 'react-router-scroll/lib/useScroll'
-import { NotFoundView } from './not_found_view'
-import { Container } from './container'
+import { Routes } from './routes'
 import { buildEnvironment } from '../relay'
 import { events } from 'shared/events'
-import { local } from 'shared/storage'
-import { searchRoute, billRoute, aboutRoute, faqRoute, adminRoute, adminBillsRoute } from '../scenes'
-
-let hasEnteredSearch = false
-
-export function reset () {
-  hasEnteredSearch = false
-}
 
 export class AppRouter extends Component {
+  routes: React$Element<*>
+
   state: { environment: Object } = {
     environment: buildEnvironment()
+  }
+
+  constructor () {
+    super()
+    this.routes = Routes()
   }
 
   // events
   didUpdateEnvironment = (header: string) => {
     const environment = buildEnvironment({ 'Authorization': header })
     this.setState({ environment })
-  }
-
-  didChangeRoute = (route: { location: { pathname: string } }) => {
-    const { pathname } = route.location
-
-    if (!hasEnteredSearch && pathname === '/') {
-      hasEnteredSearch = true
-    }
-
-    if (hasEnteredSearch && pathname !== '/') {
-      local.set('@@legislated/intro-visited', 'true')
-    }
   }
 
   // lifecycle
@@ -50,20 +36,10 @@ export class AppRouter extends Component {
 
   render () {
     const { environment } = this.state
-    debugger
     const middleware = applyRouterMiddleware(useRelay, useScroll())
 
     return <Router history={browserHistory} render={middleware} environment={environment}>
-      <Route component={Container} onChange={this.didChangeRoute}>
-        <Route path='/' {...searchRoute} />
-        <Route path='about' {...aboutRoute} />
-        <Route path='faq' {...faqRoute} />
-        <Route path='bill/:id' {...billRoute} />
-        <Route path='admin' {...adminRoute}>
-          <IndexRoute {...adminBillsRoute} />
-        </Route>
-        <Route path='*' component={NotFoundView} />
-      </Route>
+      {this.routes}
     </Router>
   }
 }
